requirements.txt
pandas
pytorch-lightning
hydra-core
torch
dvc
